Ae realiza un cambio en el archivo src/app/(backoffice)/dashboard/tracking/page.tsx para agregar un nuevo componente de mapa que muestra la ubicación de los operadores en tiempo real.
